requirements.txt
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
fonttools==4.28.5
gunicorn==20.1.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
kaleido==0.2.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
prompt-toolkit==3.0.24
pycodestyle==2.8.0
Pygments==2.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
six==1.16.0
tenacity==8.0.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
